requirements.txt
# Kern Resources - Python Dependencies

# ============================================================================
# Core AI and ML
# ============================================================================
crewai==0.121.1
openai>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.5.0
groq>=0.8.0

# ============================================================================
# Web Framework and GUI
# ============================================================================
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
chainlit>=2.5.5
streamlit>=1.28.0

# ============================================================================
# Memory and Vector Database
# ============================================================================
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
faiss-cpu>=1.7.0

# ============================================================================
# Data Processing
# ============================================================================
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# ============================================================================
# Environment and Configuration
# ============================================================================
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0.0

# ============================================================================
# Development and Testing
# ============================================================================
pytest>=7.4.0
pytest-flask>=1.2.0
black>=23.0.0
flake8>=6.0.0

# ============================================================================
# Deployment and Packaging
# ============================================================================
pyinstaller>=5.13.0
auto-py-to-exe>=2.40.0
gunicorn>=21.0.0

# ============================================================================
# Utilities
# ============================================================================
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0
pathlib2>=2.3.0
